Created first draft of solving engine, previous move functionality
Made solver engine class, methods to explore state space of a given game
with aim of finding a target configuration. Used a greedy search,
however the metric that is being used to score boards is not ideal:
isn't sensitive to changes made outside of the best scoring subgrid.

Added in functionality to the game grid class to track the move that was
made to get to a child board.